refactor: reingeniería del CORE para soporte multi-sucursal y sincronización offline
1.Estructura Multi-sede: Creación del modelo Sucursal y migración de la tabla Empleados para incluir relación foránea obligatoria por sucursal.
2.Identificadores Globales: Implementación de factura_local_uuid en pedidos para garantizar la unicidad de transacciones generadas fuera de línea.
3.Trazabilidad de Líneas: Adición de detalle_local_uuid en detalles de pedido para mantener la integridad referencial con la base de datos local.
4.Control de Sincronización: Inclusión de sellos de tiempo ultima_modificacion en productos e inventarios para gestionar descargas de datos incrementales.
5.Esquemas de Idempotencia: Refactorización de PedidoCreate para permitir la recepción de UUIDs y evitar la duplicidad de registros en reintentos de red.
6.Seguridad Distribuida: Modificación del esquema de EmpleadoResponse para incluir el hash de contraseña necesario para validaciones locales.
7.Contextualización de Sesión: Extensión del login para devolver sucursal_id, permitiendo a la aplicación cliente filtrar datos por ubicación automáticamente.
8.Módulo de Catálogos: Despliegue del router sync_catalogos para la distribución eficiente de productos, precios e impuestos hacia las sucursales.
9.Módulo de Usuarios: Implementación del router sync_usuarios para la actualización de perfiles autorizados por sede en la base de datos local.
10.Lógica de Persistencia: Ajuste del endpoint de pedidos para procesar y almacenar metadatos de sincronización provenientes de la capa de integración.
11.Gestión de Stock: Actualización de la función de inventario para forzar el refresco de timestamps globales tras cada venta o movimiento de mercancía.
12.Compatibilidad SQL Server: Configuración de tipos UNIQUEIDENTIFIER de SQLAlchemy para asegurar la persistencia correcta de UUIDs en el motor MSSQL.

+++ app/schemas/reportes_schema.py
@@ -18,11 +18,10 @@ class RankingProductosResponse(BaseModel):
     class Config:
         from_attributes = True
 
-
 class AlertaStockResponse(BaseModel):
     nombre: str
-    cantidad_disponible: Decimal
-    stock_minimo: Decimal
+    cantidad_disponible: int
+    stock_minimo: int
 
     class Config:
         from_attributes = True
